implemented ambient & temp sensors: working

+++ libraries/AP_Temp_I2C/AP_Temp_I2C.h
@@ -29,7 +29,7 @@ protected:
 private:
 	void _timer(void);
 	void acquireTemp(void);
-	uint16_t _temperature;
+	float _temperature;
 	//uint16_t _ambient_temperature;
 	AP_HAL::OwnPtr<AP_HAL::I2CDevice> _dev;
